REFACTOR: extract additional factors for women
